Got rid of all compiler warnings
The g++ compiler was throwing warnings about the absence of a virtual
destructor in polymorphic classes that are derived from 'Menu' class, so
I added the virtual destructor...
And there was 'Wreorder' warnings when initilizing const variables in
the constructor of 'MotiveCreature', so I reordered them manually.
And finally some int to unsigned int comparision warnings.
*NOTE: the 'breakPoint' variable is intended to be there just to use gdb lol.