Finished working on 'Logger' class
Class Logger is now ready and functioning properly, and I've used it
to detect a leak in the creation and deletion of EnemyBlock objects in
EnemiesManager.
The bug reasons is not specified yet, I'll continue to replace
std::set<Enemy*> with std::forward_list<Enemy*> inside EnemiesManager
and see what happens then, it will be easier to handle and less error
prone...
*NOTE: this version of the code compiles successfully only when 'DEBUG'
macro is defined, I'll fix this in the next commit.