Replace 'std::set' with 'std::forward_list' as enemies container
now the container of enemies objects in EnemiesManager class is of
type 'std::forward_list <std::unique_ptr<Enemy>>', this made the
code much safer and easier to read and reduced the risk of memory
leak when adding/removing enemies of different types.
The bug detected in a previous commit about a memory leak in enemies
container is now solved with this change.
*Note: the game balance is still sick tho, but I'll make this a beta
release and adjust the balance later for the stable release.