Added sound effects to button clicks and some other modifications.
- Added new class 'AudioManager' which is responsible for loading
and playing sound/music files properly, for now it only plays
button click sound.
- Renamed some functions that were accidently named with this
convention: <functionName> to this convention: <function_name>
because it contradicts with naming conventions specified in the
Readme.md file.

*DETECTED POSSIBLE BUG: the game runs smoothly at 60FPS on my integrated
graphics card, but when I run it on NVIDIA MX110 it runs at 58FPS
causing some glitches in motion of some objects (Player and CannonBall).