More enhancement for game balance.
-	now 'warning' rectangles are drawn when an enemy is outside
the screen in a corner, for example `warningDown` and `warningRight`
might be drawn at the same time if an enemy get outside of the
screen from the buttom right corner of the screen.
(there's a small bug with this feature which is that nothing is
drawn when the enemy is outside from the upper left corner, I
couldn't figure out why it didn't work...).
-	EnemyBlockSettings was modified a little bit to make `medium`
difficulty less difficult.