Attempt to fix the bug reported by Ali Ibrahem
The game ran at 120FPS on Ali's PC but he was unable to move the player
at all, I suspected that the issue is in the 'fix precision errors'
section in `Player::update` method because it is causing small movement
to turn into 0.
So, I tried to fix it, but can't fully confirm it IS fixed until Ali
test the new version on his machine.